Move ShaderHelpView to help package.
